requirements.txt
Pillow==6.2.1
requests==2.20.0